server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
email-validator
passlib
bcrypt==3.2.0
python-jose[cryptography]
python-multipart
